add folder_as_output class attribute  converters
The obj is to know if a converter without writer will save to folder or to a file. So, the right data is send or an error is thrown based on the file reolution from the user data.

it is set to none. All converters without writer should set it. While it is default to False for others